Only check for mismatched (de)alloc in C++
C++ can use any of `{*alloc,new,new[]}`, each of which requires matching
deallocator. C, on the other hand, can only use `*alloc` to allocate and
all of the pointers obtained such way are disposed of with `free`.
Therefore it should not be needed to run the check on C code.

Note: the statement above dismisses existence of such functions as
`VirtualAlloc`, `mmap`, `sbrk`, etc., but they were not part of this
check anyway.