C#: Revert recent change to `AccessorCall`
The recent change to `AccessorCall` on dd9952556624571c4f66a5e4ffe5a5155cf772e6 resulted
in some bad join-orders, so I have (partly) reverted them. This means that the issues
orignally addressed by that change are now reintroduced, and I plan to instead apply a
fix to the CFG, which--unlike the original fix--should be able to handle multi-property-tuple
assignments.