C++: Rewrite RedundantNullCheckSimple.ql for perf
The performance of this query appears to have regressed since I first
tested it on 5 February 2019. On Wireshark I was getting this summary of
slowest predicates:

    #select#cpe#13#ff ................................... 49m44s
    #select#cpe#13#ff#antijoin_rhs ...................... 6m10s
    #select#cpe#13#ff#shared ............................ 4m9s
    Element::ElementBase::toString_dispred#ff ........... 37s
    Instruction::Instruction::toString_dispred#ff ....... 29.2s
    IRBlock::IRBlock::dominates_dispred#ff .............. 21.3s
    IRBlock::IRBlock::dominates_dispred#ff_10#join_rhs .. 9.1s

The main cause of slowness is a CPE predicate inserted by the optimiser,
but the query wasn't likely to be fast in the first place because it
used `dominates`, which is a huge relation, in a 3-way join for which
there was no good join order.

I've rewritten the query with manual recursion to follow the edges of
the `immediatelyDominates` relation, stopping as early as possible. This
gives the following summary of slowest predicates:

    Element::ElementBase::toString_dispred#ff ........... 36.3s
    Instruction::Instruction::toString_dispred#ff ....... 31.1s