C++: Fix SSAConstruction::hasPhiNode performance
This was the slowest single predicate in the Wireshark IR generation.
The slowness was caused by a bad join order in some iterations. For
example:

    (1249s) Starting to evaluate predicate SSAConstruction::Cached::hasPhiNode#ff#cur_delta
    (1317s) Tuple counts:
    166166    ~2%      {2} r1 = SCAN SSAConstruction::Cached::hasDefinition#fff#prev_delta OUTPUT FIELDS {SSAConstruction::Cached::hasDefinition#fff#prev_delta.<0>,SSAConstruction::Cached::hasDefinition#fff#prev_delta.<1>}
    726101769 ~0%      {3} r2 = JOIN r1 WITH SSAConstruction::Cached::variableLiveOnEntryToBlock#ff ON r1.<0>=SSAConstruction::Cached::variableLiveOnEntryToBlock#ff.<0> OUTPUT FIELDS {r1.<1>,SSAConstruction::Cached::variableLiveOnEntryToBlock#ff.<1>,r1.<0>}
    82911     ~53%     {2} r3 = JOIN r2 WITH Dominance::getDominanceFrontier#ff ON r2.<0>=Dominance::getDominanceFrontier#ff.<0> AND r2.<1>=Dominance::getDominanceFrontier#ff.<1> OUTPUT FIELDS {r2.<2>,r2.<1>}
    10467     ~33%     {2} r4 = r3 AND NOT SSAConstruction::Cached::hasPhiNode#ff#prev(r3.<0>,r3.<1>)
                       return r4
    (1317s) 			 - SSAConstruction::Cached::hasPhiNode#ff_delta has 8185 rows (order for disjuncts: delta=10000).

As the above RA shows, `hasPhiNode#ff` used to be a join of three
relations, and two of those were not part of the recursion. To fix
performance, a helper predicate now joins those two relations in
advance, adding some context to keep the size of the helper predicate
small. The helper predicate has about twice as many rows as there are
basic blocks on Wireshark.

With these changes, `hasPhiNode` in `aliased_ssa` drops from 1m9s to
being invisible in the clause timing report. The helper predicate,
`getFrontierBlockIfDefinedIn`, takes about 2s.

Since I needed to get all definitions of a virtual variable, I also
cleaned up the use of the predicate that computes those definitions.