C++: Use the IR from the LGTM suite
To keep performance in check, this PR also disables the most expensive
hidden-by-default queries. See CPP-371. I've tried to keep the hidden
queries that are not overly expensive and are security-related.

The `SignedOverflowCheck.ql` and `PointerOverflow.ql` queries are
disabled because they use the old `GlobalValueNumbering` library, which
is slow.