C++: Data flow through & and *
The data flow library conflates pointers and their objects in some
places but not others. It might be ideal to avoid that conflation, but
that's not realistic without using the IR.

We've had good experience in the taint tracking library with conflating
pointers and objects, and it improves results for field flow, so perhaps
it's time to try it out for all data flow.