Python: Use ExternalStringKind instead of UntrustedStringKind
This is mostly to bring consistency between the other web libraries. But with
the following reasoning, it is also TheRightThing™

ExternalStringKind is an abstract base class. Currently it has concrete
subclasses UntrustedStringKind and ExternalXmlString.

Before we were doing this

    override predicate isSourceOf(TaintKind kind) {
        kind instanceof UntrustedStringKind
    }

However, if the unstrusted string is deserialized as XML, we will not report the
possible vulnerability (since this requires sink to have kind
ExternalXmlString). By using ExternalStringKind in isSourceOf, we would be
able to report that vulnerability (yay!).