C++: Downgrade ExprHasNoEffect to medium precision
We keep adding special cases to this query, but a quick glance over its
results on lgtm.com shows that we have many more special cases to cover.

The fundamental problem with this query is that it doesn't know where
the encapsulation barriers are. As a consequence, the query will flag
calls like `destroy_my_object(&my_object);` when the `destroy_my_object`
function has no effect _in the particular version and configuration_
we've extracted.