Modernizes ExpectedMappingForFormatString.ql
The ObjectAPI.qll file was modified because the original Objects API (which is,
confusingly, not in ObjectAPI.qll) had predicates for sequence types, map types,
etc. The modification ports some of that over to make the modernized query work.

The expected results for the query seem to have been incorrect in the previous
version of the query that used the Objects API instead of the Values API. It
looked to be the case that the single result in the new expected file was the
intended result, given that expected_mapping_for_fmt_string has a list argument
to a mappy format string. The expected file has therefore also been updated.