Adds modernizations for TruncatedDivision.ql
Some work still needs to be done to make it testable. The previous version
doesn't seem to have had a test associated with it, and the canonical reference
example doesn't work immediately, possibly due to version issues. The query only
applies to Python 2, but qltest might be treating the example as Python 3.