Adds test for TruncatedQuery.
Unfortunately, this query doesn't seem to actually do what it is described as
doing. PointsTo analysis (as well as RefersTo before it) don't have values for
function calls like `sum(xs)` or `len(ys)`, it seems, and so the query can only
find the most obvious cases of truncation where the value origins are clearly
numeric. This is despite the original query's example being precisely one
involving `sum` and `len`.