Divide and conquer broken due to overflow (CWE-190)
Divide and conquer algorithms like binary search or merge sort
have to calculate a mid point, when they are dividing the task.
(Splitting the task into [0, mid] and [mid, arrayLength])
This is often done as mid = (low+high)/2, but this will fail, when
low and high are big numbers (2^30). In this case overflow will have happened,
leading to a negative value.
Using this negative value to access an array will then lead to an
ArrayIndexOutOfBoundsException.