Python: Fix inverted `subscriptUnknown` predicate.
This predicate is used to indicate whether we're likely to have full information
about subscripted values of the given kind of object. Thus, it doesn't hold for,
say, `SequenceObjectInternal`, but it _does_ hold for, say,
`UnknownInstanceInternal`.

For `BuiltinOpaqueObjectInternal`, we look at whether the underlying built-in
has items that can be accessed. Unfortunately, for whatever reason this check
was inverted: If there _were_ such items, we would say that subscripting was
unknown, and if there _weren't_, we would say that subscripting gave a known
result. The upshot of this was that subscripting would never work on these
objects.

Negating the body of the predicate was not a valid solution, as this would
induce a negative recursion, so I have split the getting of the built-in into
its own positive step, and the negation is only on the `getItem` call, which is
fine (it calls directly into the underlying extensional predicate).

I don't know whether anything has been affected by this bug -- most built-ins
aren't "opaque" by our definition, so probably this won't change much in terms
of results.