C++: Fix IR generation for implicit non-void returns
While debugging some other IR issues, I noticed that the sanity queries were failing on the existing IR dataflow tests. It turns out that this was due to those tests having functions that return `int`, but don't actually have a `return` statement. Executing past the end of a function body without returning a value is undefined behavior, but we should still generate well-formed IR.

The fix is surprisingly simple. Since we already treated `return e;` pretty much the same as the initialization of a variable, all we had to do was treat plain `return;` as if it were a variable initialization with no initializer. This lets the existing code just "initialize" the variable via an `Uninitialized` instruction.

I added an IR test case. The changes to the `syntax-zoo` test expectations are due to the fact that we now generate different rubbish IR when there are ODR violations, not because of any real change.