Python: BottleRoute.getRoute => .getUrlPattern
writing documentation for this made me realize that `getRoute` on a class that
is called route is a bit confusing. `getUrlPattern` should be much more to the point!

I originally picked the route name to be consistent with django, but I'm really happy
with `getUrlPattern`, and would like to use that for other libraries as well
down the line.