Python: Rename test directories to have non-package names.
In general, we cannot expect overlapping modules to be disambiguated
based on the _packages_ they reside in (as they may be inside
directories that cannot be interpreted as packages). To better model
this, I have changed the names from `copy1` to `copy-1` etc.

This also cleans up the test case a bit. Note that it still contains
twice as many results as we want.