Python: Add `Module::importsFrom` and use it to prevent overlap.
Basically, this predicate looks for a common `Container` that contains
both the module itself and the module being imported. In general there
will be many such containers (since if a given container has this
property, then so does its parent), so we require that the container
itself is "close enough" to properly disambiguate the imported module
based solely on its name.