C++: Don't rely on reverse reads to obtain nested store steps, as this creates bad performance. In the AST dataflow library the 'final' reverse read step will read from the beginning of a chain of field dereferences (i.e., the x in x.y.z = source()). Typically, this start node will not be a source in the localFlowStep relation. This is different in the IR, where the 'final' reverse read step (up until this commit) was the total chi operand, and this total chi operand will typically be the source of several simpleLocalFlowStep results. This can create pathological cases where a node can both end a big step in the dataflow library, but also be part of a _bigger_ big step. This caused a huge performance regression on openjdk.
This commit avoids reverse reads, and instead creates explicit store
steps for nested field dereferences. We compute the final step from the
FieldNode to the total chi operand as a simpleFlowStep instead, thus
avoiding the performance problem.

Lesson learned: Be sure that the target of your readStep isn't also a
source in your localFlowStep relation.