Python: Add example of top-level module shadowing stdlib
Although this test is added under the `wrong` folder, the current results from
this CodeQL test is actually correct (compared with the Python
interpreter). However, they don't match what the extractor does when invoked
with `codeql database create`. I want to use this test-case as a reference, when
updating the tool invocation from our testing setup to match what happens in the
extractor.

It would be better to just fix the extractor, but that is not trivial, since it
requires us to correctly guess the root of source-code in a project (so we can
simulate it becoming the first element of `sys.path`). Making the behavior of
our testing setup the same as the production environment is a step in the right
direction in my opinion (even if that means having the same known problems).

Inspired by the debugging in https://github.com/github/codeql/issues/4640