Java/Android : Insecure Context Creation
Using untrusted input to create a context can lead to arbitary code execution
This PR adds a query to detect cases where an untrusted input may be
used to instantiate a new context in the application's scope. Allowing
this results in arbitary code execution as a malicious app would now be
able to take control of the execution flow.