C#: Remove duplicate named type population
Named types might be populated multiple times due to `NullableAnnotation` being part of the symbol. For example, reference types can be extracted max. 3-times: for annotated, non-annotated, and oblivious (such as the declaration) nullable contexts. Ideally, we should populate types exactly once. If the declaration is from the source code, we can limit the processing to the oblivious context without any additional memory footprint. If the symbol is defined outside the current assembly, we're still populating all variants.