Java: Denial of Service due to decoding of untrusted input
Decoding of variables can throw exceptions. When these exceptions are
not caught, these can cause the applicaiton to creash leading to a
denial of service. Best example of this vulnerability is [CVE-2020-13933](https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2020-13933) which occured in Apache Shiro. In this case, the decoding of untrusted cookie
values lead to a applciation crash which lead to a denial of service.